#171 use WriteType in flows
