+++ packages/flows-plugin/src/Flows/types.ts
@@ -1,16 +1,14 @@
-import { FormattedValueFragmentType } from '@neuint/term-js';
+import { WriteType } from '@general/types/write';
 
 export type StepResultType = {
   to?: string,
-  data?: string | FormattedValueFragmentType,
-  duration?: number,
-  withSubmit?: boolean,
+  write?: WriteType,
 };
 
 export type FlowType = Array<{
   autostart?: boolean,
   onEnter?: (data: { [key: string]: string }) => void;
-  write?: { data: string | FormattedValueFragmentType, duration?: number, withSubmit?: boolean };
+  write?: WriteType;
   variableName?: string;
   secret?: boolean;
   handler?: (data: { [key: string]: string }) => Promise<StepResultType | undefined>;
